Cleaned up init.lua
